examples/agent/pyproject.toml
[tool.poetry]
name = "avis-agent"
version = "2024.01.3"
description = "Agent to acquire pictures and interact with AVIS."
authors = ["Lucas Vandroux <lucas.vandroux@vu.engineering>"]
readme = "README.md"
packages = [{include = "avis_agent"}]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
pillow = "^10.0.0"
urllib3 = "^2.0.7"
certifi = "^2023.7.22"
python-dateutil = "^2.8.2"
types-pillow = "^10.0.0.2"
opencv-python = "^4.9.0.80"
aiohttp = "^3.8.3"
numpy = "^1.25.2"
# Pinned to 3.3.2 because of https://github.com/pymodbus-dev/pymodbus/issues/1711
# as soon as this is resolved upstream we can remove the pinning
pymodbus = {version = "3.3.2", extras = ["repl"]}
pydantic = "^2.1.1"
pydantic-settings = "^2.0.3"
pytz = "^2023.3"
frozendict = "^2.3.8"
tenacity = "^8.2.3"
click = "^8.1.7"
# TODO(apagano): fix this
# vue_instrumentation = { path = "../../libs/vue-instrumentation", develop = true }
# vue_avis_client = { path = "../../libs/vue-avis-client", develop = true }

[tool.poetry.group.luxonis]
optional = true

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.luxonis.dependencies]
depthai = "^2.17.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-datadir = "^1.4.1"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple"
priority = "supplemental"

[tool.poetry.scripts]
avis_agent = 'avis_agent.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
